Fotos y RecyclerView funcionando en nuevo Recuerdo. Implementado swipe para borrar

+++ app/src/main/java/com/agp/mybox/Modelo/DAO/RecursoDAO.java
@@ -23,8 +23,8 @@ public interface RecursoDAO {
     public void borrarRecurso(Recurso recurso);
 
     @Query("SELECT * FROM recurso WHERE id = :id")
-    public LiveData<List<Recurso>> listarRecurso(long id);
+    public LiveData<List<Recurso>> listarRecurso(int id);
 
     @Query("SELECT * FROM recurso WHERE idRecuerdo = :id")
-    public LiveData<List<Recurso>> RecursosDeRecuerdo(long id);
+    public LiveData<List<Recurso>> RecursosDeRecuerdo(int id);
 }
